requirements.txt
Adafruit-Blinka==8.20.1
adafruit-circuitpython-busdevice==5.2.6
adafruit-circuitpython-neopixel==6.3.9
adafruit-circuitpython-pixelbuf==2.0.2
adafruit-circuitpython-requests==2.0.1
adafruit-circuitpython-typing==1.9.4
Adafruit-PlatformDetect==3.49.0
Adafruit-PureIO==1.1.11
asn1crypto==0.24.0
certifi==2018.8.24
chardet==3.0.4
colorzero==1.1
cryptography==2.6.1
devscripts==2.19.5+deb10u1
entrypoints==0.3
gpiozero==1.6.2
idna==2.6
keyring==17.1.1
keyrings.alt==3.1.1
meson==0.56.1
paho-mqtt==1.6.1
pycrypto==2.6.1
pyftdi==0.54.0
PyGObject==3.30.4
PySDL2==0.9.3
pyserial==3.5
python-apt==1.8.4.3
pyudev==0.21.0
pyusb==1.2.1
pyxdg==0.25
requests==2.21.0
rpi-ws281x==5.0.0
RPi.GPIO==0.7.0
SecretStorage==2.3.1
six==1.12.0
spidev==3.5
ssh-import-id==5.7
typing-extensions==4.7.1
urllib3==1.24.1
urwid==2.0.1
